+++ Krowi_AchievementFilter/Data/ExportedCalendarEvents.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-11-28 23-04-26 ]] --
+-- [[ Exported at 2021-11-30 19-43-59 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --